Remove SimperbyApi trait
